DynaFlow Builder/requirements.txt
customtkinter==5.1.2
darkdetect==0.8.0
Pillow==9.4.0
pure-python-adb==0.3.0.dev0